Refactored ReceiveMoveCharacter()
