+++ Source Main 5.2/source/WSclient.cpp
@@ -9,7 +9,6 @@
 #include "ZzzInventory.h"
 #include "ZzzLodTerrain.h"
 #include "ZzzAI.h"
-#include "ZzzPath.h"
 #include "ZzzTexture.h"
 #include "ZzzEffect.h"
 #include "ZzzOpenglUtil.h"
@@ -1760,94 +1759,108 @@ void ReceiveMoveCharacter(std::span<const BYTE> ReceiveBuffer)
         return;
     }
 
-    auto pathData = const_cast<BYTE*>(ReceiveBuffer.subspan(fixedSize).data());
-
-    int  Key = ((int)(Data->KeyH) << 8) + Data->KeyL;
+    int Key = ((int)(Data->KeyH) << 8) + Data->KeyL;
     CHARACTER* c = &CharactersClient[FindCharacterIndex(Key)];
     OBJECT* o = &c->Object;
-    PATH_t* pCharPath = &c->Path;
+
+    if (c->Dead || !o->Live)
+    {
+        return;
+    }
+
+    if (IsMonster(c))
+    {
+        MUHelper::g_MuHelper.AddTarget(Key, false);
+    }
 
     c->PositionX = Data->SourceX;
     c->PositionY = Data->SourceY;
     c->TargetX = Data->TargetX;
     c->TargetY = Data->TargetY;
+    c->TargetAngle = Data->PathMetadata >> 4;
 
-    if (c->Dead == 0 && o->Live)
+    g_ConsoleDebug->Write(MCD_RECEIVE, L"ID : %s | sX : %d | sY : %d | tX : %d | tY : %d", c->ID, c->PositionX, c->PositionY, c->TargetX, c->TargetY);
+
+    if (Key == HeroKey)
     {
-        c->TargetAngle = Data->PathMetadata >> 4;
-        if (Key == HeroKey)
+        if (!c->Movement)
         {
-            if (!c->Movement)
-            {
-                c->PositionX = Data->TargetX;
-                c->PositionY = Data->TargetY;
-            }
+            c->PositionX = Data->TargetX;
+            c->PositionY = Data->TargetY;
         }
-        else
-        {
-            pCharPath->Lock.lock();
-
-            if (c->Movement)
-            {
-                c->Movement = false;
-                SetPlayerStop(c);
-            }
-
-            if (o->Type == MODEL_CRUST)
-            {
-                c->PositionX = Data->TargetX;
-                c->PositionY = Data->TargetY;
-                o->Angle[2] = CreateAngle(o->Position[0], o->Position[1], c->PositionX * 100.f, c->PositionY * 100.f);
-                SetPlayerWalk(c);
-                c->JumpTime = 1;
-            }
-            else if (c->Appear == 0)
-            {
-                int iDefaultWall = TW_CHARACTER;
-
-                if (gMapManager.WorldActive >= WD_65DOPPLEGANGER1 && gMapManager.WorldActive <= WD_68DOPPLEGANGER4
-                    && Key != HeroKey)
-                {
-                    iDefaultWall = TW_NOMOVE;
-                }
-
-                BYTE direction = c->TargetAngle;
-                POINT nextPos = MovePoint(static_cast<EPathDirection>(direction), { Data->SourceX, Data->SourceY });
-
-                pCharPath->PathX[0] = nextPos.x;
-                pCharPath->PathY[0] = nextPos.y;
-
-                for (int i = 0; i < pathNum; ++i)
-                {
-                    int byteIndex = i / 2;
-
-                    direction = (i % 2 == 0) ?
-                        (pathData[byteIndex] >> 4) & 0x0F :
-                        (pathData[byteIndex] & 0x0F);
-
-                    nextPos = MovePoint(static_cast<EPathDirection>(direction), nextPos);
-
-                    pCharPath->PathX[i + 1] = nextPos.x;
-                    pCharPath->PathY[i + 1] = nextPos.y;
-                }
-
-                pCharPath->PathNum = pathNum + 1;
-                pCharPath->Direction = direction;
-                pCharPath->CurrentPath = 0;
-                pCharPath->CurrentPathFloat = 0;
-            }
-
-            pCharPath->Lock.unlock();
-
-            if (IsMonster(c))
-            {
-                MUHelper::g_MuHelper.AddTarget(Key, false);
-            }
-
-            g_ConsoleDebug->Write(MCD_RECEIVE, L"ID : %s | sX : %d | sY : %d | tX : %d | tY : %d", c->ID, c->PositionX, c->PositionY, c->TargetX, c->TargetY);
-        }
+        return;
     }
-    
+
+    if (o->Type == MODEL_CRUST)
+    {
+        c->PositionX = Data->TargetX;
+        c->PositionY = Data->TargetY;
+        o->Angle[2] = CreateAngle(o->Position[0], o->Position[1], c->PositionX * 100.f, c->PositionY * 100.f);
+        SetPlayerWalk(c);
+        c->JumpTime = 1;
+
+        return;
+    }
+
+    c->Movement = false;
+    SetPlayerStop(c);
+
+    auto pathData = const_cast<BYTE*>(ReceiveBuffer.subspan(fixedSize).data());
+
+    if (pathData == nullptr || pathNum == 0)
+    {
+        c->PositionX = Data->TargetX;
+        c->PositionY = Data->TargetY;
+
+        return;
+    }
+
+    PATH_t* pCharPath = &c->Path;
+    pCharPath->Lock.lock();
+
+    if (c->Appear == 0)
+    {
+        int iDefaultWall = TW_CHARACTER;
+
+        if (gMapManager.WorldActive >= WD_65DOPPLEGANGER1 && gMapManager.WorldActive <= WD_68DOPPLEGANGER4
+            && Key != HeroKey)
+        {
+            iDefaultWall = TW_NOMOVE;
+        }
+
+        POINT nextPos = { Data->SourceX, Data->SourceY };
+
+        for (int i = 0; i < pathNum; ++i)
+        {
+            int byteIndex = i / 2;
+
+            BYTE direction = (i % 2 == 0) ?
+                (pathData[byteIndex] >> 4) & 0x0F :
+                (pathData[byteIndex] & 0x0F);
+
+            switch (static_cast<EPathDirection>(direction))
+            {
+            case EPathDirection::WEST:       nextPos.x--; nextPos.y--; break;
+            case EPathDirection::SOUTHWEST:  nextPos.y--; break;
+            case EPathDirection::SOUTH:      nextPos.x++; nextPos.y--; break;
+            case EPathDirection::SOUTHEAST:  nextPos.x++; break;
+            case EPathDirection::EAST:       nextPos.x++; nextPos.y++; break;
+            case EPathDirection::NORTHEAST:  nextPos.y++; break;
+            case EPathDirection::NORTH:      nextPos.x--; nextPos.y++; break;
+            case EPathDirection::NORTHWEST:  nextPos.x--; break;
+            }
+
+            pCharPath->PathX[i] = nextPos.x;
+            pCharPath->PathY[i] = nextPos.y;
+        }
+
+        pCharPath->PathNum = pathNum;
+        pCharPath->Direction = c->TargetAngle;
+        pCharPath->CurrentPath = 0;
+        pCharPath->CurrentPathFloat = 0;
+    }
+
+    pCharPath->Lock.unlock();
 }
 
 void ReceiveMovePosition(const BYTE* ReceiveBuffer)
